+++ package.json
@@ -34,11 +34,11 @@
     "fastify": "4.17.0",
     "fastify-healthcheck": "^4.4.0",
     "mongodb": "^5.5.0",
-    "winston": "^3.3.3",
-    "winston-daily-rotate-file": "^4.5.5",
     "morgan": "^1.10.0",
     "node-cache": "^5.1.2",
-    "node-cleanup": "^2.1.2"
+    "node-cleanup": "^2.1.2",
+    "winston": "^3.3.3",
+    "winston-daily-rotate-file": "^4.5.5"
   },
   "devDependencies": {
     "@chainsafe/eslint-config": "^1.1.0",
@@ -47,24 +47,24 @@
     "@types/cors": "^2.8.12",
     "@types/eslint": "^8",
     "@types/jest": "^27.0.1",
+    "@types/mongodb": "^4.0.7",
     "@types/morgan": "^1.9.3",
     "@types/node": "^16.7.10",
-    "@types/supertest": "^2.0.11",
-    "@types/mongodb": "^4.0.7",
     "@types/node-cleanup": "^2.1.2",
+    "@types/supertest": "^2.0.11",
     "cross-fetch": "3.1.5",
+    "docker-secret": "^1.2.4",
+    "dotenv": "^10.0.0",
     "dotenv-cli": "^4.0.0",
+    "dotenv-flow": "^3.2.0",
     "eslint": "^8.21.0",
     "jest": "^27.1.0",
     "nodemon": "^2.0.12",
     "prettier": "^2.7.1",
     "prisma": "^4.1.0",
     "supertest": "^6.1.6",
-    "ts-jest": "29.1.0",
+    "ts-jest": "^29.1.1",
     "ts-node": "^10.2.1",
-    "typescript": "^4.7.4",
-    "docker-secret": "^1.2.4",
-    "dotenv": "^10.0.0",
-    "dotenv-flow": "^3.2.0"
+    "typescript": "^4.7.4"
   }
 }
